modal arrow shows only when needed
