+++ FrontEnd/assets/style.css
@@ -370,6 +370,7 @@ footer nav ul {
 	font-size: 16px;
 	/* margin-bottom: 16px; */
 }
+
 .modal__body {
 	flex: 1;
 	box-sizing: border-box;
