+++ FrontEnd/index.html
@@ -15,7 +15,7 @@
 	<aside id="modal" class="modal" aria-hidden="true" role="dialog" aria-modal="false" aria-labelledby="modal__title" style="display: none;">
 		<div class="modal-wrapper js-modal-stop">
 			<div class="modal__header">
-				<a href="#modal" data-version="gallery" class="js-modal material-symbols-outlined">arrow_back</a>
+				<a href="#modal" data-version="gallery" class="js-modal material-symbols-outlined modal__arrow">arrow_back</a>
 				<button class="js-modal-close material-symbols-outlined">close</button>
 			</div>
 			<div class="modal__body">
